+++ test/Main.hs
@@ -56,14 +56,14 @@ tests = testGroup "Bytes"
         @=?
         Bytes.stripOptionalSuffix (bytes "h") (bytes "hey man")
     ]
-  , testGroup "sharedPrefix"
+  , testGroup "longestCommonPrefix"
     [ testProperty "finds prefix" $ \(pre :: Bytes) (a :: Bytes) (b :: Bytes) ->
         if | Just (wa,_) <- Bytes.uncons a, Just (wb,_) <- Bytes.uncons b, wa /= wb ->
-               Bytes.sharedPrefix (pre <> a) (pre <> b) === pre
+               Bytes.longestCommonPrefix (pre <> a) (pre <> b) === pre
            | otherwise -> property Discard
     , testProperty "finds no prefix" $ \(a :: Bytes) (b :: Bytes) ->
         if | Just (wa,_) <- Bytes.uncons a, Just (wb,_) <- Bytes.uncons b, wa /= wb ->
-               Bytes.sharedPrefix a b === mempty
+               Bytes.longestCommonPrefix a b === mempty
            | otherwise -> property Discard
     ]
   , testGroup "dropWhileEnd"